pyproject.toml
[tool.poetry]
name = "yoflow"
version = "1.0.2"
description = "Django workflows"
authors = ["Yoyo <dev@yoyowallet.com>"]
include = [
    "LICENSE",
]
packages = [
    { include="yoflow" },
]

[tool.poetry.dependencies]
python = "^3.9"

django = ">=3.2,<5.0"
djangorestframework = "*"
six = "*"
psycopg2-binary = "^2.9.10"

[tool.poetry.group.dev.dependencies]
pytest-mock = "*"
pytest-django = "*"
pytest = "*"
factory-boy = "*"
xenon = "*"
flake8 = "*"
pytest-cov = "*"
pre-commit = "*"
ruff = "^0.11.0"

[tool.black]
include = '\.pyi?$'
skip-string-normalization = true
exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE='tests.settings'

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
